+++ app/routes/games/search.js
@@ -1,24 +1,19 @@
 var router = require('express').Router();
 var Game = require('../../models/game');
-var GameArchive = require('../../models/game-archive');
 
 router.post('/', function(req, res) {
   var searches;
   if (req.body.slug) {
     searches = [
       Game.getBySlug(req.body.slug).select('slug'),
-      GameArchive.getBySlug(req.body.slug).select('slug')
     ];
   } else if (req.body.bundleId) {
     searches = [
       Game.getByBundleId(req.body.bundleId).select('slug'),
-
-      GameArchive.getByBundleId(req.body.bundleId).select('slug')
     ];
   } else if (req.body.search) {
     searches = [
       Game.getBySearch(req.body.search, 10),
-      GameArchive.getBySearch(req.body.search, 10)
     ];
   }
 
@@ -29,6 +24,7 @@ router.post('/', function(req, res) {
         data[0][i].url = '/games/game/';
       }
     }
+    //TODO: Get fix for this from other branch
     if (null !== data[0]) {
       for (var n = 0; n < data[1].length; n++) {
         data[1][n].url = '/archive/game/';
