--- app/views/archive/game.jade
@@ -1,133 +1,0 @@
-extend ../templates/archive
-
-append gameContent
-	.panel.panel-default
-		.panel-heading
-			if isEditor
-				include modal-edit
-				button.pull-right.btn.btn-default.btn-sm(data-toggle="modal" data-target="#editGame")
-					span.glyphicon.glyphicon-pencil
-					|  Edit
-			h3.panel-title Properties
-		.panel-body
-			.row
-				.col-sm-2: strong Location
-				.col-sm-10: p
-					a(href=game.location)=game.location
-			.row
-				.col-sm-2: strong Repository
-				.col-sm-10: p
-					a(href=game.repository)=game.repository
-			if game.description
-				.row
-					.col-sm-2: strong Description
-					.col-sm-10: p=game.description
-			.row
-				.col-sm-2: strong Capabilities
-				.col-sm-10: .well.clearfix
-					.col-sm-4.capabilities
-						p: strong UI
-						ul.list-unstyled
-							li
-								span.glyphicon(class=game.capabilities.ui.mouse ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Mouse
-							li
-								span.glyphicon(class=game.capabilities.ui.touch ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Touch
-					.col-sm-4.capabilities
-						p: strong Sizes
-						ul.list-unstyled
-							li
-								span.glyphicon(class=game.capabilities.sizes.xsmall ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Extra-Small
-							li
-								span.glyphicon(class=game.capabilities.sizes.small ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Small
-							li
-								span.glyphicon(class=game.capabilities.sizes.medium ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Medium
-							li
-								span.glyphicon(class=game.capabilities.sizes.large ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Large
-							li
-								span.glyphicon(class=game.capabilities.sizes.xlarge ? "glyphicon-ok yes" : "glyphicon-remove no")
-								|  Extra-Large
-					.col-sm-4.capabilities
-						p: strong Features
-						ul.list-unstyled
-							if game.capabilities.features.webgl
-								li
-									span.glyphicon.glyphicon-ok.yes
-									|  WebGL
-							if game.capabilities.features.webaudio
-								li
-									span.glyphicon.glyphicon-ok.yes
-									|  WebAudio API
-							if game.capabilities.features.webworkers
-								li
-									span.glyphicon.glyphicon-ok.yes
-									|  Web Worker API
-							if game.capabilities.features.websockets
-								li
-									span.glyphicon.glyphicon-ok.yes
-									|  WebSockets API
-							if game.capabilities.features.geolocation
-								li
-									span.glyphicon.glyphicon-ok.yes
-									|  Geolocation
-			.row
-				.col-sm-2: strong Created
-				.col-sm-10: p.help-inline=moment(game.created).fromNow()
-			.row
-				.col-sm-2: strong Updated
-				.col-sm-10: p.help-inline=moment(game.updated).fromNow()
-
-	.panel.panel-default
-		.panel-heading
-			h3.panel-title Deployments
-		.panel-body
-			.row.deployment
-				.col-sm-2: strong Production
-				.col-sm-10
-					.input-group
-						input.monospace.form-control.select-all(readonly type="text" value='<iframe src="//#{host}/embed/#{game.slug}" frameborder="0" />')
-						.input-group-btn
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?controls=1&title=1")
-								span.glyphicon.glyphicon-play
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?debug=1&controls=1&title=1")
-								span.glyphicon.glyphicon-wrench
-								span.glyphicon.glyphicon-play
-			.row.deployment
-				.col-sm-2: strong Staging
-				.col-sm-10
-					.input-group
-						input.monospace.form-control.select-all(readonly type="text" value='<iframe src="//#{host}/embed/#{game.slug}?status=stage&token=#{token}" frameborder="0" />')
-						.input-group-btn
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=stage&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-play
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=stage&debug=1&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-wrench
-								span.glyphicon.glyphicon-play
-			.row.deployment
-				.col-sm-2: strong QA
-				.col-sm-10
-					.input-group
-						input.monospace.form-control.select-all(readonly type="text" value='<iframe src="//#{host}/embed/#{game.slug}?status=qa&token=#{token}" frameborder="0" />')
-						.input-group-btn
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=qa&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-play
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=qa&debug=1&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-wrench
-								span.glyphicon.glyphicon-play
-
-			.row.deployment
-				.col-sm-2: strong Development
-				.col-sm-10
-					.input-group
-						input.monospace.form-control.select-all(readonly type="text" value='<iframe src="//#{host}/embed/#{game.slug}?status=dev&token=#{token}" frameborder="0" />')
-						.input-group-btn
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=dev&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-play
-							a.btn.btn-default(href="//#{host}/embed/#{game.slug}?status=dev&debug=1&controls=1&title=1&token=#{token}")
-								span.glyphicon.glyphicon-wrench
-								span.glyphicon.glyphicon-play
