--- app/views/archive/index.jade
@@ -1,29 +1,0 @@
-extend ../templates/normal
-
-append title
-	| Edit Game -
-
-append content
-	.col-sm-12: .panel.panel-default
-		.panel-heading
-			.pull-right
-				span Sort by:
-				.btn-group.sort-group
-					a.btn.btn-sm.btn-default(class=(order=='alphabetical' ? 'disabled' : '') href="/archive/alphabetical") Title
-					a.btn.btn-sm.btn-default(class=(order=='latest' ? 'disabled' : '') href="/archive/latest") Updated
-			h3.panel-title Games
-		.panel-body
-			if !games.length
-				.alert.alert-warning.text-center No games have been archived.
-			else
-				.row
-					for game in games
-						.col-md-2.col-sm-3.col-xs-6
-							a.game(data-toggle="tooltip" title=game.title href="/archive/game/#{game.slug}")
-								if game.thumbnail && game.thumbnail.length
-									img.preview(src="data:image/png;base64,#{game.thumbnail}")
-								else
-									img.preview(src="/images/none.png")
-									span.title=game.title
-									span.badge=game.releases.length
-				.row: include ../partials/pagination
