--- app/views/archive/modal-edit.jade
@@ -1,46 +1,0 @@
-.modal.fade#editGame
-	.modal-dialog
-		form.modal-content.form-horizontal(method="post" action=url)
-			.modal-header
-				button.close(data-dismiss="modal") &times;
-				h4.modal-title Edit Game Properties
-			.modal-body
-				.form-group
-					label.col-sm-3.control-label(for="title") Title
-					.col-sm-9
-						input.form-control#title(type="text" name="title" #{disabled} value=game.title required)
-				if isAdmin
-					.form-group(data-unique="/games/search" data-ignore=game.slug)
-						label.col-sm-3.control-label(for="slug") Slug
-						.col-sm-9
-							input.form-control#slug(type="text" name="slug" #{disabled} value=game.slug required)
-							span.yes.glyphicon.glyphicon-ok.form-control-feedback
-							span.no.glyphicon.glyphicon-remove.form-control-feedback
-				.form-group
-					label.col-sm-3.control-label(for="repository") Repository
-					.col-sm-9
-						input.form-control#repository(type="text" #{disabled} name="repository" value=game.repository required)
-				.form-group
-					label.col-sm-3.control-label(for="location") Location
-					.col-sm-9
-						input.form-control#location(type="text" #{disabled} name="location" value=game.location required)
-				.form-group
-					label.col-sm-3.control-label(for="thumbnail") Thumbnail
-					.col-sm-9
-						.base64(class=(game.thumbnail && game.thumbnail.length ? "" : "empty") data-limit="60000" data-width="200" data-height="100")
-							button.select.btn.btn-default.btn-sm(type="button" data-toggle="popover" data-trigger="manual" data-content="Invalid image") Browse...
-							input.input.hidden(type="file" accept="image/jpeg, image/png")
-							button.btn.btn-default.btn-sm.reset(type="button") &times;
-							img.preview(width="200" height="100" src="data:image/png;base64,#{game.thumbnail}")
-							input.output(type="hidden" name="thumbnail" value=game.thumbnail)
-							span.help.help-inline PNG or JPG 200px &times; 100px less than 60KB
-				.form-group
-					label.col-sm-3.control-label(for="description") Description
-					.col-sm-9
-						textarea.md.form-control#description(name="description" #{disabled})=game.description
-				include capabilities
-			.modal-footer
-				button.btn.btn-primary(type="submit" name="action" value="updateGame") Update
-				button.btn.btn-primary(data-toggle="confirm" type="submit" name="action" value="restoreGame") Restore
-				button.btn.btn-danger(data-toggle="confirm" type="submit" name="action" value="removeGame") Delete
-				button.btn.btn-default(type="button" data-dismiss="modal") Cancel
