--- app/views/archive/release.jade
@@ -1,47 +1,0 @@
-extend ../templates/archive
-
-append gameContent
-
-	.panel.panel-default
-		.panel-heading
-			h3.panel-title Edit Release
-		form.form.form-horizontal.panel-body(method="post" action=url)
-			.form-group
-				label.col-sm-3.control-label(for="commitId") Commit ID
-				.col-sm-9
-					input.form-control#commitId(type="text" name="commitId" value=release.commitId required)
-			.form-group
-				label.col-sm-3.control-label(for="version") Version
-				.col-sm-9
-					input.form-control#version(type="text" name="version" value=release.version placeholder="1.0.0")
-			.form-group
-				label.col-sm-3.control-label(for="status") Status
-				.col-sm-9
-					table.table.table-bordered.table-striped.no-margin.text-center
-						thead: tr
-							th.text-center: label(for="statusDev") Dev
-							th.text-center: label(for="statusQA") QA
-							th.text-center: label(for="statusStage") Stage
-							if isAdmin
-								th.text-center: label(for="statusProd") Prod
-						tbody: tr
-							td: input#statusDev(type="radio" name="status" value="dev" checked=(release.status == "dev"))
-							td: input#statusQA(type="radio" name="status" value="qa" checked=(release.status == "qa"))
-							td: input#statusStage(type="radio" name="status" value="stage" checked=(release.status == "stage"))
-							if isAdmin
-								td: input#statusProd(type="radio" name="status" value="prod" checked=(release.status == "prod"))
-			.form-group
-				label.col-sm-3.control-label(for="notes") Notes
-				.col-sm-9
-					textarea.autogrow.form-control.md#notes(name="notes")=release.notes
-					p.help-block Notes input supports
-						a(href="https://en.wikipedia.org/wiki/Markdown") markdown
-						| syntax.
-			include capabilities
-			.form-group
-				.col-sm-9.col-sm-offset-3.text-center
-					input(type="hidden" name="updatedBy" value=user._id)
-					input(type="hidden" name="release" value=release._id)
-					button.btn.btn-lg.btn-primary(type="submit" name="action" value="updateRelease") Update
-					button.btn.btn-lg.btn-danger(data-toggle="confirm" type="submit" name="action" value="removeRelease") Delete
-					a.btn.btn-lg.btn-default(href="#{url}s") Cancel
