--- app/views/templates/archive.jade
@@ -1,22 +1,0 @@
-extend ../templates/normal
-
-append title
-	| #{game.title} - Game -
-
-append content
-	.col-sm-3: .row
-		if game.thumbnail && game.thumbnail.length
-			.col-xs-6.col-sm-12
-				img.btn-block.preview(src="data:image/png;base64,#{game.thumbnail}")
-
-		.col-sm-12(class=(game.thumbnail && game.thumbnail.length ? "col-xs-6" : "")): .list-group
-			a.list-group-item(href='/archive/game/#{game.slug}' class=isActive('/archive/game/#{game.slug}')) Home
-			a.list-group-item(href='/archive/game/#{game.slug}/releases' class=isActive(/releases?$/)) Releases
-				span.badge=game.releases.length
-			if isAdmin
-				a.list-group-item(href='/archive/game/#{game.slug}/privileges' class=isActive(/privileges$/)) Privileges
-	.col-sm-9
-		h2=game.title
-		if isEditor
-			include ../partials/feedback
-		block gameContent
