+++ package.json
@@ -29,6 +29,7 @@
     "jade": "^1.9.2",
     "jquery": "^3.0.0",
     "marky-markdown": "^12.0.0",
+    "method-override": "3.0.0",
     "moment": "^2.9.0",
     "mongoose": "~4.0.4",
     "mongoose-deep-populate": "^1.0.1",
